test(json): split unit and integration tests
